rename: image repo to image storage
