+++ Titinski.WebAPI/Interfaces/Storage/IImageStorage.cs
@@ -1,8 +1,8 @@
 ﻿using Titinski.WebAPI.Models;
 
-namespace Titinski.WebAPI.Services.ImageRepository
+namespace Titinski.WebAPI.Interfaces.Storage
 {
-    public interface IImageRepo
+    public interface IImageStorage
     {
         /// <summary>
         /// Adds the image included in the request
@@ -10,6 +10,6 @@ namespace Titinski.WebAPI.Services.ImageRepository
         /// <param name="rant">The Rant object received from client</param>
         /// <returns>Image URI in the repo</returns>
         public string AddRant(RantPost rant);
-        public Rant GetRant(string id);
+        public Rant GetRant(string path);
     }
 }
